Board now fetches its data and allows drawing
Added server-side board component for fetching its data and forwarding to the client-side board.
The client board allows drawing however it is not saved anywhere.
Added a slight grey background when hovering over clickable elements.
The cursor changes to a pencil when drawing but needs fixing — it's way too large.